+++ tests/conftest.py
@@ -7,7 +7,7 @@ from pytest import fixture
 @fixture
 def protocol_parser():
     """Basic instance of ProtocolParser"""
-    from dcspy.dcs.bios import ProtocolParser
+    from dcspy.dcsbios import ProtocolParser
     return ProtocolParser()
 
 
@@ -20,7 +20,7 @@ def lcd_size() -> Tuple[int, int]:
     :return: width and height
     :rtype: Tuple[int, int]
     """
-    from dcspy.utils.lcd_sdk import MONO_WIDTH, MONO_HEIGHT
+    from dcspy.lcd_sdk import MONO_WIDTH, MONO_HEIGHT
     return MONO_WIDTH, MONO_HEIGHT
 
 
@@ -32,7 +32,7 @@ def hornet(lcd_size: Tuple[int, int]):
     :return: F/A-18C Hornet instance
     :rtype: FA18Chornet
     """
-    from dcspy.dcs.aircrafts import FA18Chornet
+    from dcspy.aircrafts import FA18Chornet
     return FA18Chornet(*lcd_size)
 
 
@@ -44,7 +44,7 @@ def black_shark(lcd_size: Tuple[int, int]):
     :return: Ka-50 Black Shark instance
     :rtype: Ka50
     """
-    from dcspy.dcs.aircrafts import Ka50
+    from dcspy.aircrafts import Ka50
     return Ka50(*lcd_size)
 
 
@@ -56,5 +56,5 @@ def tomcat(lcd_size: Tuple[int, int]):
     :return: F-14B Tomcat instance
     :rtype: F14B
     """
-    from dcspy.dcs.aircrafts import F14B
+    from dcspy.aircrafts import F14B
     return F14B(*lcd_size)
