+++ tests/test_lcd_sdk.py
@@ -13,7 +13,7 @@ from pytest import mark
                                              ('logi_lcd_color_set_title', ('', (1, 2, 3)), False),
                                              ('logi_lcd_color_set_text', (1, '', (1, 2, 3)), False)])
 def test_all_failure_cases(function, args, result):
-    from dcspy.utils import lcd_sdk
+    from dcspy import lcd_sdk
     lcd_sdk.LCD_DLL = None
     assert getattr(lcd_sdk, function)(*args) is result
 
@@ -29,7 +29,7 @@ def test_all_failure_cases(function, args, result):
                                                     ('logi_lcd_color_set_title', 'LogiLcdColorSetTitle', ('', (1, 2, 3)), True),
                                                     ('logi_lcd_color_set_text', 'LogiLcdColorSetText', (1, '', (1, 2, 3)), True)])
 def test_all_success_cases(py_func, c_func, args, result):
-    from dcspy.utils import lcd_sdk
+    from dcspy import lcd_sdk
     mocked_c_func = Mock()
     mocked_c_func.return_value = result
     lcd_sdk.LCD_DLL = {c_func: mocked_c_func}
@@ -39,7 +39,7 @@ def test_all_success_cases(py_func, c_func, args, result):
 @mark.parametrize('c_func, effect, lcd, size', [('logi_lcd_mono_set_background', [True], 1, (16, 4)),
                                                 ('logi_lcd_color_set_background', [False, True], 2, (32, 24))])
 def test_update_display(c_func, effect, lcd, size):
-    from dcspy.utils import lcd_sdk
+    from dcspy import lcd_sdk
     from PIL import Image
     with patch.object(lcd_sdk, 'logi_lcd_is_connected', side_effect=effect) as connected:
         with patch.object(lcd_sdk, c_func, return_value=True) as set_background:
@@ -57,7 +57,7 @@ def test_update_display(c_func, effect, lcd, size):
                                                         [call(0, ''), call(1, ''), call(2, ''), call(3, ''),
                                                          call(4, ''), call(5, ''), call(6, ''), call(7, '')])])
 def test_clear_display(c_funcs, effect, lcd, size, text):
-    from dcspy.utils import lcd_sdk
+    from dcspy import lcd_sdk
     with patch.object(lcd_sdk, 'logi_lcd_is_connected', side_effect=effect) as connected:
         with patch.object(lcd_sdk, c_funcs[0], return_value=True) as set_background:
             with patch.object(lcd_sdk, c_funcs[1], return_value=True) as set_text:
